created wrap_wasm
