+++ rust/Cargo.toml
@@ -3,8 +3,9 @@ members = [
     "wrap_schema/tree_sitter_wrap", 
     "wrap_schema/reim_schema",
     "reim_dt",
-    "reim_wrap", 
     "reim_dt_wasm",
+    "reim_wrap", 
+    "reim_wrap_wasm", 
     "reim_codegen", 
     "reim_cli", 
     "reim_cli",
