+++ rust/app/Cargo.toml
@@ -7,3 +7,8 @@ edition = "2021"
 reim_wrap = { path = "../reim_wrap" }
 reim_dt_wasm = { path = "../reim_dt_wasm" }
 reim_codegen = { path = "../reim_codegen" }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0.85"
+async-trait = "0.1.58"
+by_address = "1.1.0"
+lazy_static = "1.4.0"
