+++ rust/app/src/wrappers/external_references/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2021"
 
 [dependencies]
 reim_wrap = { path = "../../../../reim_wrap" }
+reim_wrap_wasm = { path = "../../../../reim_wrap_wasm" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.85"
 async-trait = "0.1.58"
