--- rust/app/src/wrappers/external_references/src/polywrap/dt/mod.rs
@@ -1,4 +1,0 @@
-pub mod imports;
-pub mod main;
-pub mod malloc;
-pub use main::send;
