+++ rust/app/src/wrappers/external_references/src/polywrap/external/module/host_wrap_module.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 use async_trait::async_trait;
 use reim_wrap::ExternalModule;
-use crate::polywrap::dt::send;
+use reim_wrap_wasm::send;
 
 pub struct HostWrapModule;
