+++ rust/app/tests/mock_test.rs
@@ -1,9 +1,14 @@
+mod app;
+
 fn main() {
 }
 
 #[cfg(test)]
 mod tests {
+    use crate::app::*;
+
     #[test]
     fn test() {
+        testReceiveReference();
     }
 }
