+++ rust/reim_wrap/src/external_module.rs
