+++ rust/reim_wrap/src/lib.rs
@@ -1,7 +1,7 @@
 #![feature(async_closure)]
 
-pub mod internal_wrap_module;
-pub mod external_wrap_module;
-pub use internal_wrap_module::*;
-pub use external_wrap_module::*;
+pub mod internal_module;
+pub mod external_module;
+pub use internal_module::*;
+pub use external_module::*;
 pub mod wrapper_module;
