+++ rust/reim_wrap_wasm/src/lib.rs
@@ -1,0 +1,7 @@
+pub mod send;
+pub mod receive;
+pub mod imports;
+pub mod malloc;
+pub mod abort;
+pub use send::*;
+pub use receive::*;
