+++ rust/reim_wrap_wasm/src/malloc.rs
